Add grouping by customer address and organize a bit
