+++ client/src/components/Reports/Customer/GeneralCustomerResultPage.js
@@ -1,19 +1,20 @@
 import React from "react";
 import { Container, Table } from "reactstrap";
-import { Spinner, Title, BackButton } from "../index";
+import { Spinner, Title, BackButton } from "../../index";
 import PropTypes from "prop-types";
-import useFetch from "../../hooks/useFetch";
-import ResultTableHead from "./ResultTableHead";
-import GeneralBranchTable from "./GeneralBranchTable";
+import useFetch from "../../../hooks/useFetch";
+import ResultTableHead from "../ResultTableHead";
+import GeneralCustomerTable from "./GeneralCustomerTable";
 
-const GeneralBranchResultPage = ({
+const GeneralCustomerResultPage = ({
 	start_date,
 	finish_date,
 	state,
 	setState,
+	type,
 }) => {
 	const { data, error, isLoading } = useFetch(
-		`/general_reports/branch/${state.customer_id}/${start_date}/${finish_date}`
+		`/general_reports/customer/${start_date}/${finish_date}`
 	);
 
 	if (error) {
@@ -23,7 +24,11 @@ const GeneralBranchResultPage = ({
 	} else if (data) {
 		return (
 			<Container>
-				<Title text={`General report of ${state.customer} addresses`} />
+				<Title
+					text={`General report${
+						type === "customer" ? "" : " for addresses"
+					} of customers`}
+				/>
 				<h3 className="text-center mt-4 mt-md-5 mb-5 mb-md-5">
 					{"Work duration from " + start_date + " to " + finish_date}
 				</h3>
@@ -32,30 +37,32 @@ const GeneralBranchResultPage = ({
 				) : (
 					<Table striped hover responsive>
 						<ResultTableHead
-							labels={["Address", "Planned duration", "Actual duration"]}
+							labels={["Customer", "Planned duration", "Actual duration"]}
 							detailed={false}
 						/>
-						<GeneralBranchTable
+						<GeneralCustomerTable
 							data={data.rows}
 							state={state}
 							setState={setState}
+							type={type}
 						/>
-						<GeneralBranchTable data={data.totals} tableFooter={true} />
+						<GeneralCustomerTable data={data.totals} tableFooter={true} />
 					</Table>
 				)}
 				<div className="d-flex justify-content-end mt-5">
-					<BackButton state={state} setState={setState} />
+					<BackButton state={state} setState={setState} type={type} />
 				</div>
 			</Container>
 		);
 	}
 };
 
-GeneralBranchResultPage.propTypes = {
+GeneralCustomerResultPage.propTypes = {
 	start_date: PropTypes.string,
 	finish_date: PropTypes.string,
+	type: PropTypes.string,
 	state: PropTypes.object,
 	setState: PropTypes.func,
 };
 
-export default GeneralBranchResultPage;
+export default GeneralCustomerResultPage;
