+++ client/src/components/Reports/General/GeneralReport.js
@@ -1,8 +1,8 @@
 import React, { useContext } from "react";
 import { useHistory } from "react-router-dom";
-import { Form, Button } from "reactstrap";
-import SelectDateU from "./SelectDateU";
-import { GeneralReportContext } from "../../contexts/GeneralReport";
+import { Form, Button, Label, Input } from "reactstrap";
+import SelectDateU from "../SelectDateU";
+import { GeneralReportContext } from "../../../contexts/GeneralReport";
 
 const CreateGeneralReport = () => {
 	const [state, setState] = useContext(GeneralReportContext);
@@ -11,6 +11,11 @@ const CreateGeneralReport = () => {
 	const handleSubmit = (e) => {
 		e.preventDefault();
 		history.push("/general_results"); // Go to <GeneralReportResultPage>
+	};
+
+	const handleChange = (e) => {
+		const { name, checked } = e.target;
+		setState({ ...state, [name]: checked });
 	};
 
 	return (
@@ -29,8 +34,22 @@ const CreateGeneralReport = () => {
 					attributeTitle="Finish date"
 				/>
 			</div>
-			<div className="d-flex justify-content-end">
-				<Button>Run</Button>
+			<div className="d-flex justify-content-between justify-content-sm-end flex-sm-column">
+				<Label
+					check
+					className="d-flex align-items-center user-select-none pl-4 text-right align-self-sm-end mb-sm-3 mb-md-4 mb-lg-5"
+					size="lg"
+				>
+					<Input
+						name="byCustomer"
+						type="checkbox"
+						onChange={handleChange}
+						checked={state.byCustomer}
+						className="mb-1"
+					/>
+					Group by customer
+				</Label>
+				<Button className="align-self-sm-end">Run</Button>
 			</div>
 		</Form>
 	);
