+++ client/src/components/Reports/ResultPage.js
@@ -22,7 +22,7 @@ const ResultPage = ({ start_date, finish_date, detailed, name, id, type }) => {
 			<Container>
 				<Title text={name} />
 				<h3 className="text-center mt-4 mt-md-5 mb-5 mb-md-5">
-					{"Work duration from " + start_date + " to " + finish_date}
+					Work duration from {start_date} to {finish_date}
 				</h3>
 				{data.rows.length < 1 ? (
 					<p>No data for this period.</p>
