+++ client/src/components/index.js
@@ -31,10 +31,10 @@ export { default as WorkerFilter } from "./Recurring/WorkerFilter";
 export { default as WorkerJobsList } from "./WorkerJobs/WorkerJobsList";
 export { default as WorkerLogTimeForm } from "./WorkerJobs/WorkerLogTimeForm";
 export { default as WorkerJobInfo } from "./WorkerJobs/WorkerJobInfo";
-export { default as CreateCustomerReport } from "./Reports/CreateCustomerReport";
-export { default as CreateWorkerReport } from "./Reports/CreateWorkerReport";
-export { default as CreateBranchReport } from "./Reports/CreateBranchReport";
-export { default as GeneralReport } from "./Reports/GeneralReport";
-export { default as GeneralWorkerTable } from "./Reports/GeneralWorkerTable";
-export { default as GeneralWorkerResultPage } from "./Reports/GeneralWorkerResultPage";
+export { default as CreateCustomerReport } from "./Reports/Customer/CreateCustomerReport";
+export { default as CreateWorkerReport } from "./Reports/Worker/CreateWorkerReport";
+export { default as CreateBranchReport } from "./Reports/Branch/CreateBranchReport";
+export { default as GeneralReport } from "./Reports/General/GeneralReport";
+export { default as GeneralWorkerTable } from "./Reports/Worker/GeneralWorkerTable";
+export { default as GeneralWorkerResultPage } from "./Reports/Worker/GeneralWorkerResultPage";
 export { default as HomepageText } from "./Homepage/HomepageText";
