+++ client/src/pages/BranchReportPage.js
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
 import { Redirect } from "react-router-dom";
 import { BranchReportContext } from "../contexts/BranchReport";
-import BranchResultPage from "../components/Reports/BranchResultPage";
+import BranchResultPage from "../components/Reports/Branch/BranchResultPage";
 
 const BranchReportPage = () => {
 	const [state] = useContext(BranchReportContext);
